ci: Use nix-env to install tools
This is easier to use instead of having to go through nix-shell for each
command. This also gets rid of the nix-shell in the shebang for formatters.sh
so that we don't require nix-shell for its use. Different versions of tools
will be caught by CI anyway so its fine to be lenient for users.