Properly substitute command path in Run() error
If the command executed by (*command).Run() exits with a nonzero status,
an error is returned which includes the complete command that was
executed, including the resolved path to the command binary. Or that's
the intention. What it actually did was prepend the resolved path to the
command + arguments, chopping off the first byte of the latter. As a
result, a call to zfs("list") which would run "/usr/bin/zfs list" would,
if the command failed, return an error claiming that "/usr/bin/zfs
fs list" was executed! This can mislead users into thinking that the
failed command was due to the wrong arguments being passed, when the
problem lies elsewhere.

Change (*command).Run() to return an error containing the resolved path
and arguments of the failed command without any spurious arguments, and
to debug-log the command with the resolved path.

Signed-off-by: Cory Snider <csnider@mirantis.com>